requirements.txt
flask==1.1.2
flask-cors==3.0.10
stripe==2.55.2
slackclient==2.9.3